requirements.txt
fastapi==0.112.0
uvicorn==0.23.2
pydantic==2.6.1
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
boto3==1.34.0